requirements.txt
Django
django-constance==3.1.0
django-picklefield==3.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
redis==5.0.3